Add a second window to show a slideshow of the picture list
The window has custom chrome which fades into the background after
a few seconds. It reappears when the mouse moves over the top of
the window. The window can be resized and dragged to a new location.

Left and right keyboard keys move to the previous and next pictures
in the picture list.