Fix play button visibility
For playing media from the media list, the play button was collapsed
and the pause button was visible when a video was selected after the
media window was already visible.

Now the set of controls that are visible when the media window is
visible is completely separate from the set of controls that are
visible when it is not. This separation makes the calculation of
the visibility state much simpler.