Show a message when import finds no media files
Only media files are added to the media list. This means that all
files in a directory can be dropped onto the UI and it will just
work. When none of the files are media files, however, the fact
that nothing happens gives the appearance that something is failing.
This message will help the user understand that nothing failed.