Simplify to display on a secondary screen
Previous versions of the app had two different display modes. One
was to drive content on a primary monitor and display on a secondary
monitor (e.g. TV). The other was to display in a small window on
the primary monitor. The idea of the second method was to make it
possible to share content during a video conference call.

Now only the first method is used. Content will always display
on the secondary monitor.