Simplify main window startup positioning
Previous builds made it possible to select one of four starting
locations: upper-left, upper-right, lower-right, and lower-left.

Now the app starts in the upper-left of the primary display the
first time the app is run. While it is running, its location is stored
whenever the window moves. The starting position of the window the next
time the app opens will be the most recently saved location.