Fix a regression
After changing the layout logic to store the current location of
the main window so that it restores at next program startup, it
would load minimized if the program ended while the main window
was in a minimized state.

Now the main window will load at the (0, 0) coordinate position
when it closes in a minimized state.