Make the positioning of the media window more flexible
The logic used until this commit has worked when a secondary monitor
is set up as the viewing monitor. It may be, however, that more than
two monitors are used. In this case the current logic fails to place
the media window in the correct location.

It would be ideal to present to the user a list of monitors so that
one could be selected as the display monitor. However, the Windows
API does not provide an exact bounding rectangle for each monitor
it lists.

The logic added in this commit is a compromise. It allows the user
to select the location of the media window by presenting a sample
window which the user then places in the correct location. The
bounding rectangle of that window is then stored and reused when
showing the actual media window.